Fix validatePartitionMetadataConsistency and improve tests
