+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/admin/AdminApiTest.java
@@ -2916,8 +2916,6 @@ public class AdminApiTest extends MockedPulsarServiceBaseTest {
         publishMessagesOnPersistentTopic("persistent://prop-xyz/ns1/ds2-partition-2", 0);
         assertEquals(admin.topics().getList("prop-xyz/ns1"),
                 List.of("persistent://prop-xyz/ns1/ds2-partition-2"));
-        // Create a topic
-        admin.topics().createPartitionedTopic("persistent://prop-xyz/ns1/ds2", 3);
 
         // create consumer and subscription
         @Cleanup
